Moved the validator to Post.js and deleted the Model.js file and its folder.
